% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooths.R
\name{bandwidth_imbens}
\alias{bandwidth_imbens}
\title{Choose bandwidth for regression discontintuity
based on Imben's Criterion}
\usage{
bandwidth_imbens(rv, y)
}
\arguments{
\item{rv}{running variable}

\item{y}{outcome variable}

\item{cutoff}{cutoff(s) for discontinuity}
}
\description{
Choose bandwidth for regression discontintuity
based on Imben's Criterion
}
