% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{bin}
\alias{bin}
\title{Bin a variable}
\usage{
bin(rv, binwidth = NULL, method = "mccrary", signif_digits = 7)
}
\arguments{
\item{rv}{running variable to put in bins}

\item{binwidth}{bin width to use for binning procedure}

\item{method}{method to use for automatic binning if bw is null}
}
\description{
Bin a variable
}
\details{
If bw (the bandwidth parameter) is null,
rv (the running variable) will be placed into bins based on
the automated procedure determined by the method argument.
Options currently supported are "mccrary" and "freedman".
for details see \link{bw_freedman} and \link{bw_mccrary}.
}
