% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{bw_freedman}
\alias{bw_freedman}
\title{Freedman-Diaconis Histogram Binning Rule}
\usage{
bw_freedman(x)
}
\arguments{
\item{x}{vector to bin}
}
\description{
Freedman-Diaconis Histogram Binning Rule
}
\references{
Freedman, David, and Persi Diaconis. "On the histogram as a density estimator:
\ifelse{html}{\out{L<sub>2</sub>} theory."
Zeitschrift fur Wahrscheinlichkeitstheorie und verwandte Gebiete 57.4 (1981): 453-476.
}
