% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{bw_mccrary}
\alias{bw_mccrary}
\title{McCrary Bandwidth Estimator for Histogram Bins}
\usage{
bw_mccrary(x)
}
\arguments{
\item{x}{vector to bin}
}
\description{
McCrary Bandwidth Estimator for Histogram Bins
}
\references{
McCrary, Justin. "Manipulation of the running variable in the
regression discontinuity design: A density test."
Journal of econometrics 142.2 (2008): 698-714.
}
