% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooths.R
\name{calc_mccrary_rule}
\alias{calc_mccrary_rule}
\title{Calculate mccrary density rule}
\usage{
calc_mccrary_rule(x, y, cutoff)
}
\arguments{
\item{x}{x variable, bin midpoints from \link{make_grid}}

\item{y}{y variable, frequency of observations in those groups}

\item{cutoff}{cutoff to be used for the rdd design}
}
\description{
Calculate mccrary density rule
}
