% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density-plot.R
\name{density_plot}
\alias{density_plot}
\title{Plot McCrary Density Test}
\usage{
density_plot(
  data,
  running_variable,
  cutoff,
  binwidth = NULL,
  title = NULL,
  subtitle = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{data}{data to be used by plot}

\item{running_variable}{Running variable for regression
discontinuity design}

\item{cutoff}{Cutoff to be used as the discontinuity}

\item{binwidth}{Bin width to be used in histogram,
will guess via [guess_binwidth] if none is specified}
}
\description{
Plot McCrary Density Test
}
\details{
Plots a histogram with a vertical line at the
discontinuity. Useful to check for manipulation.
}
\examples{
# produces a density plot
density_plot(hansen_dwi, "bac1", cutoff = c(0.08))
}
