% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-regression.R
\name{getResponse}
\alias{getResponse}
\title{Get response variable from formula}
\usage{
getResponse(formula)
}
\arguments{
\item{formula}{formula to be used}
}
\description{
Get response variable from formula
}
