% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooths.R
\name{guess_bandwidth}
\alias{guess_bandwidth}
\title{Estimate bandwidth for regression discontinuity}
\usage{
guess_bandwidth(rv, cutoff, y = NULL, method)
}
\arguments{
\item{rv}{running variable}

\item{cutoff}{cutoff(s) for discontinuity}

\item{y}{outcome variable}
}
\description{
Estimate bandwidth for regression discontinuity
}
\details{
For the final bandwidth for the regression discontinuity
design, you should include a y variable to choose
Imben's criteria (see \link{bandwidth_imbens}). If you are
estimating a bandwidth to be used in a sorting test,
to make sure people aren't sorting on the running variable,
you omit y entirely.
}
