% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density-plot.R
\name{guess_binwidth}
\alias{guess_binwidth}
\title{Guess binwidth for histogram}
\usage{
guess_binwidth(x)
}
\arguments{
\item{x}{vector to use in histogram}
}
\description{
Guess binwidth for histogram
}
\details{
This uses the Freedman–Diaconis rule
to determine width of bins. To be used as a default
in case the user doesn't specify their own parameter
}
