% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{guess_bw}
\alias{guess_bw}
\title{Guess binwidth for histogram}
\usage{
guess_bw(x, method = c("mccrary", "freedman"))
}
\arguments{
\item{x}{vector to use in histogram}

\item{method}{method to use for automatic binning if bw is null}
}
\description{
Guess binwidth for histogram
}
\details{
The method guesses the bandwidth to be used as a default
in case the user doesn't specify their own parameter.
Options currently supported are "mccrary" and "freedman".
for details see \link{bw_freedman} and \link{bw_mccrary}.
}
