% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{weight_gaussian}
\alias{weight_gaussian}
\alias{weight_rectangular}
\alias{weight_triangular}
\alias{weight_epanechnikov}
\alias{weight_biweight}
\alias{weight_cosine}
\alias{weight_optcosine}
\title{Available kernel weighting functions}
\usage{
weight_gaussian(x, bw)

weight_rectangular(x, bw)

weight_triangular(x, bw)

weight_epanechnikov(x, bw)

weight_biweight(x, bw)

weight_cosine(x, bw)

weight_optcosine(x, bw)
}
\arguments{
\item{x}{vector to weight}

\item{bw}{bandwidth to use}
}
\description{
Available kernel weighting functions
}
