% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\docType{data}
\name{kerns}
\alias{kerns}
\title{List of supported kernels}
\format{
An object of class \code{character} of length 7.
}
\usage{
kerns
}
\description{
List of supported kernels
}
\details{
List of supported kernel weighting schemes
}
\keyword{datasets}
