% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-regression.R
\name{localpoly}
\alias{localpoly}
\title{Given a bandwidth, run a local polynomial regression}
\usage{
localpoly(form, data, h, rv, kernel = "triangular")
}
\arguments{
\item{data}{data.frame or table to use for running regression}

\item{h}{bandwidth to use for regression}

\item{rv}{running variable to use for weighting}
}
\description{
Given a bandwidth, run a local polynomial regression
}
