% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{make_grid}
\alias{make_grid}
\title{Return Binned, Grouped Histogram for running variable,
above and below cutoff}
\usage{
make_grid(rv, cutoff, bw = NULL, normalize = T, method = "mccrary")
}
\arguments{
\item{rv}{running variable}

\item{cutoff}{cutoff to be used in discontinuity}

\item{bw}{bin width, defaults to automatic rule specified by method}

\item{normalize}{whether to normalize frequencies by the length of the vector}

\item{method}{method for determining automatic bin width, defaults to "mccrary"}
}
\description{
Return Binned, Grouped Histogram for running variable,
above and below cutoff
}
\details{
This function takes in a running variable, a vector of cutoff points,
an optional bandwidth, and an optional method for automated binning procedure.
It returns a data.table that has gone through the first stage of the McCrary density test,
e.g. one with normalized count data for the median of each bin. Bins without
any observations are dropped. For details on the default binning procedure, see
\link{bin}, \link{guess_bw}, \link{bw_mccrary}, and \link{bw_freedman}.
}
