% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate-balance.R
\name{plot_covbal}
\alias{plot_covbal}
\title{Visualize covariate balance at the discontinuity}
\usage{
plot_covbal(
  data,
  method = c("lm", "loess", "gam", "quad"),
  running_variable,
  cov,
  binwidth = NULL,
  cutoff,
  ...
)
}
\arguments{
\item{data}{Data to use for covariate balance checks}

\item{method}{Method to use for balance checks}

\item{running_variable}{running variable to use for checks}

\item{cov}{Vector of covariate names to plot}

\item{...}{other parameters to pass to function}
}
\description{
Visualize covariate balance at the discontinuity
}
\details{
The method argument defaults to a linear regression
using all the data. Other acceptable arguments are "loess"
for local linear regression, "gam" for generalized additive
models, "quadratic" for a quadratic fit using the whole
dataset.
}
