% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density-plot.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot Density Histogram}
\usage{
plot_hist(
  data,
  rv,
  cutoff,
  binwidth = NULL,
  method = "mccrary",
  title = NULL,
  subtitle = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{data}{data to be used by plot}

\item{rv}{Running variable for regression
discontinuity design}

\item{cutoff}{Cutoff to be used as the discontinuity}

\item{binwidth}{Bin width to be used in histogram,
will guess via \link{guess_binwidth} if none is specified}

\item{method}{method to be used for guessing binwidth}
}
\description{
Plot Density Histogram
}
\details{
Plots a histogram with a vertical line at the
discontinuity. Useful to check for manipulation.
}
\examples{
# produces a density plot
density_plot(hansen_dwi, "bac1", cutoff = c(0.08))
}
