% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density-plot.R
\name{plot_mccrary}
\alias{plot_mccrary}
\title{Plot for McCrary sorting test}
\usage{
plot_mccrary(
  data,
  rv,
  cutoff,
  method = NULL,
  binwidth = NULL,
  bandwidth = NULL,
  bin_method = "mccrary",
  band_method = "mccrary",
  title = NULL,
  subtitle = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{data}{data to be used by plot}

\item{rv}{name of column of running variable}

\item{cutoff}{cutoff to be tested for sorting}

\item{method}{currently ignored}

\item{binwidth}{binwidth to be used for first-stage
histogram smoother}

\item{bandwidth}{bandwidth to be used for second-stage
kernel smoother}

\item{title}{Title for the plot}

\item{subtitle}{subtitle for the plot}

\item{theme}{theme for the plot}

\item{...}{other arguments to be passed}
}
\description{
Plot for McCrary sorting test
}
