% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-rdd.R
\name{simulate_rd}
\alias{simulate_rd}
\title{Simulate regression discontinuity datasets given fitted model}
\usage{
simulate_rd(object, ...)
}
\arguments{
\item{object}{object fit with RDRobust, rdd, rdtools, or just linear regression}
}
\description{
Simulate regression discontinuity datasets given fitted model
}
\details{
This function accepts a fit that estimates a regression discontinuity,
and then assumes that the fitted model represents the true data generating
process. Then it simulates several datasets given that process, returning
them in a data.table well suited to accompanying visual checks.
}
