% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{weight}
\alias{weight}
\title{Weight according to kernel function}
\usage{
weight(x, bw, kernel)
}
\arguments{
\item{x}{vector to weight according to kernel function}

\item{bw}{bandwidth to use for weighting}

\item{kernel}{kernel weighting function to use}
}
\description{
Weight according to kernel function
}
